fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
argon2-cffi==23.1.0
pyotp==2.9.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
cryptography==43.0.1
python-dotenv==1.0.1


